# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campaign Nexus",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns/{campaign_id}/maps": {
      "get": {
        "tags": [
          "atlas"
        ],
        "summary": "List Maps",
        "operationId": "list_maps_api_v1_campaigns__campaign_id__maps_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MapSummary"
                  },
                  "title": "Response List Maps Api V1 Campaigns  Campaign Id  Maps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "atlas"
        ],
        "summary": "Upload Map",
        "operationId": "upload_map_api_v1_campaigns__campaign_id__maps_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_map_api_v1_campaigns__campaign_id__maps_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/maps/{map_id}": {
      "get": {
        "tags": [
          "atlas"
        ],
        "summary": "Get Map",
        "operationId": "get_map_api_v1_campaigns__campaign_id__maps__map_id__get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "atlas"
        ],
        "summary": "Update Map",
        "operationId": "update_map_api_v1_campaigns__campaign_id__maps__map_id__patch",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "atlas"
        ],
        "summary": "Delete Map",
        "operationId": "delete_map_api_v1_campaigns__campaign_id__maps__map_id__delete",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/maps/{map_id}/image": {
      "get": {
        "tags": [
          "atlas"
        ],
        "summary": "Get Map Image",
        "operationId": "get_map_image_api_v1_campaigns__campaign_id__maps__map_id__image_get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/maps/{map_id}/markers": {
      "post": {
        "tags": [
          "atlas"
        ],
        "summary": "Add Marker",
        "operationId": "add_marker_api_v1_campaigns__campaign_id__maps__map_id__markers_post",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/maps/{map_id}/markers/{marker_id}": {
      "patch": {
        "tags": [
          "atlas"
        ],
        "summary": "Update Marker",
        "operationId": "update_marker_api_v1_campaigns__campaign_id__maps__map_id__markers__marker_id__patch",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "marker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Marker Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "atlas"
        ],
        "summary": "Delete Marker",
        "operationId": "delete_marker_api_v1_campaigns__campaign_id__maps__map_id__markers__marker_id__delete",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "marker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Marker Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/maps/{map_id}/regions": {
      "post": {
        "tags": [
          "atlas"
        ],
        "summary": "Add Region",
        "operationId": "add_region_api_v1_campaigns__campaign_id__maps__map_id__regions_post",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/maps/{map_id}/regions/{region_id}": {
      "patch": {
        "tags": [
          "atlas"
        ],
        "summary": "Update Region",
        "operationId": "update_region_api_v1_campaigns__campaign_id__maps__map_id__regions__region_id__patch",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "atlas"
        ],
        "summary": "Delete Region",
        "operationId": "delete_region_api_v1_campaigns__campaign_id__maps__map_id__regions__region_id__delete",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          },
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_map_api_v1_campaigns__campaign_id__maps_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "map_kind": {
            "type": "string",
            "title": "Map Kind",
            "default": "region"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "parent_map_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Map Id"
          }
        },
        "type": "object",
        "required": [
          "file",
          "name"
        ],
        "title": "Body_upload_map_api_v1_campaigns__campaign_id__maps_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MapDetail": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "map_kind": {
            "type": "string",
            "title": "Map Kind"
          },
          "width_px": {
            "type": "integer",
            "title": "Width Px"
          },
          "height_px": {
            "type": "integer",
            "title": "Height Px"
          },
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "parent_map_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Map Id"
          },
          "parent_map_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Map Name"
          },
          "markers": {
            "items": {
              "$ref": "#/components/schemas/MarkerOut"
            },
            "type": "array",
            "title": "Markers"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RegionOut"
            },
            "type": "array",
            "title": "Regions"
          },
          "layers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Layers"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "name",
          "map_kind",
          "width_px",
          "height_px",
          "media_id",
          "location_id",
          "parent_map_id",
          "parent_map_name",
          "markers",
          "regions",
          "layers"
        ],
        "title": "MapDetail"
      },
      "MapSummary": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "map_kind": {
            "type": "string",
            "title": "Map Kind"
          },
          "width_px": {
            "type": "integer",
            "title": "Width Px"
          },
          "height_px": {
            "type": "integer",
            "title": "Height Px"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "parent_map_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Map Id"
          },
          "marker_count": {
            "type": "integer",
            "title": "Marker Count"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "name",
          "map_kind",
          "width_px",
          "height_px",
          "location_id",
          "parent_map_id",
          "marker_count"
        ],
        "title": "MapSummary"
      },
      "MapUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "map_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map Kind"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "parent_map_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Map Id"
          }
        },
        "type": "object",
        "title": "MapUpdate"
      },
      "MarkerCreate": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "layer": {
            "type": "string",
            "title": "Layer",
            "default": "default"
          },
          "target_entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Entity Id"
          },
          "child_map_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Map Id"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "MarkerCreate"
      },
...
      "MarkerUpdate": {
        "properties": {
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "layer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer"
          },
          "target_entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Entity Id"
          },
          "child_map_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Map Id"
          }
        },
        "type": "object",
        "title": "MarkerUpdate"
      },
...
    }
  }
}